Correctly create TopCMSItem instances with non-zero CMS counting table
Previously a TopCMSItem instance had an all-zero CMS counting table, and
we "covered" up this issue later by manually adding the instance's
"item" when constructing TopCMSInstance instances.

However, computing inner products of CMS works exclusively on the
counting table(s) of CMS instances, and here we need a proper
representation of the counting table for TopCMSItem (wrong: each row
conists of only zeroes; correct: each row is all zeroes with the
exception of one cell, which is set to 1L).