Fix flaky `MinHasherSpec` test
One `MinHasher32` being tested was using 247 hash functions internally, which
means that the expected absolute error bound:

    1 / sqrt {# of hash functions} = 1 / sqrt 247  ~  0.064

The same test required an upper bound on the error of `0.05`, which was below
the expected error bound, meaning that this test would give flaky results due to
occasionally exceeding the required error bound.  This change increases the
upper bound on the error to 0.1 which should give enough headroom over the
expected error bound to reduce the number of random test failures.