Add Batched data type.
This is the free semigroup -- it represents combining one or more
elements using an associative operation. The values are stored in a
tree structure, which can be traversed. If a semigroup for the
underlying values becomes available, the structure can be summed to a
single value.

Batching represents a space/time trade-off (we use more space to use
less time). In some cases these structures may become too big. To
address this, there are constructors for Monoid[Batched[A]] and
Semigroup[Batched[A]] instances that will periodically compact
themselves (summing the tree to a single value) to keep the overall
size below a specific parameter.

Batched is appropriate anytime an eager concatenation or combination
will produce intermediate values that will be thrown away, and a more
efficient aggregate summation is possible. In those cases, building up
a batch and then summing it should result in much more efficient
operations.

This commit adds the type, with documentation. It also adds some basic
law-checks and tests.