Improve HLLSeries performance. (#575)
* Improve HLLSeries performance.

This commit improves the performance of the HLLSeries data
structure. In particular, it introduces two new methods:

    series.insert(bytes, timestamp)
    series.approximateSizeSince(threshold)

Previously to add data to an HLLSeries required allocating another
series and then combining them with the monoid, potentially creating a
lot of garbage. The `insert` method does the same thing but without an
intermediate series.

Similarly, getting an approximate size used to require calling
`.toHLL` to build an HLL structure, and then asking that. By
abstracting out the machinery to actually do the HLL calculation,
we're able to do this directly from HLLSeries with
`approximateSizeSince`.

There are several other internal efficiency improvements. Some
independent benchmarks I've run show a 4-5x speed up in building
series' from many individual events, and a 3-4x speed up in
calculating the approximate sizes.

The data in the series and the serialization format are unchanged.

* Remove Java 8-ism.

* Add override modifier.

* Respond to review comments.

* Use .sum instead of .reduce(_ + _).

* Swap floor and ceil to widen (not narrow) bounds.

This was suggested by @johnynek.

* Respond to more review comments.