Making BloomFilters polymorphic
This makes BloomFilters polymorphic, with the addition that they require
a implicit view from A => Array[Byte] for the type that they are to
store.

It seemed there was a problem defining the implicit function inside of
the tests, which caused serialization not to work. Defining the function
outside of the actual test class and then importing it resolves this
problem. This is why the `ImplicitStringToBytesView` object is added to the
tests and then imported into the test functions.