Polymorphic Bloom filters (#607)
* Making BloomFilters polymorphic

This makes BloomFilters polymorphic, with the addition that they require
a implicit view from A => Array[Byte] for the type that they are to
store.

It seemed there was a problem defining the implicit function inside of
the tests, which caused serialization not to work. Defining the function
outside of the actual test class and then importing it resolves this
problem. This is why the `ImplicitStringToBytesView` object is added to the
tests and then imported into the test functions.

* Adding basic BloomFilter benchmarks

* Change Hash128 to require hashWithSeed

Hash128 will now require implementing classes to define `hashWithSeed` instead of `hash`. The later will now rely on using hashWithSeed with a default seed.

* BloomFilter using implicit Hash128

* Let BloomFilter benchmarks compute filter width

* Fixed typo in BloomFilter docs

* Clean up Hash128

Bringing the interfaces back the the way they previously were, and make
`DefaultSeed` a `def` rather than a `val`.

* Add return type to BloomFilter apply

* Make BloomFilter property tests generic

* Update BloomFilter docs