Fix #632 : BloomFilter saturated exception
From the paper linked to the size estimation of a bloomfilter,
we have a inegality mixing t = number bits to 1 of the filter, m=width of filter, numHash=k

        nl <= S-1(t-1)
        nr >= S-1(t+1)
S-1 is defined like :

	S-1(t) = ln(1 - t/m) / k.ln(1 - 1/m)

problem occur when _t = m_

	S-1(m) = ln(0) / k.ln(1 - 1/m)

but *ln(-1/m)* is not correct.

problem is :

	ln( 1 - 1/m) < 0
The scala maths library give :

	scala> val infinity = scala.math.log(0) / -42
	infinity: Double = Infinity

	scala> infinity.round.toInt
	res24: Int = -1

just add special case for the S-1 function.