Add DecayingCMS[K] type.
This type represents a count-min sketch whose values decay over time
according to a provided half-life.

It uses a module-style approach, where the half-life, width, depth,
and other parameters are set in a module. The actual CMS values are
path-dependent types on the DecayingCMS[K] module value. This means
that we don't have to carry around or serialize anything besides the
count-min values themselves, and also means we can be sure that CMS
values to be combined are aligned.

It's a slightly different pattern that people may be used to, but I
think it works much better for this kind of thing. (I could imagine
creating a non-decaying variant with a similar design, but that's
outside the scope of this PR.)